OpactOpt: shader refactoring

+++ modules/opactopt/glsl/opactopt/point/approximportancesum.frag
@@ -57,29 +57,20 @@ in Point {
 }
 fragment;
 
-// Opacity optimisation settings
-uniform float q;
-uniform float r;
-uniform float lambda;
-
 void main() {
     // Prevent invisible fragments from blocking other objects (e.g., depth/picking)
-    if (fragment.color.a == 0) {
-        discard;
-    }
+    if (fragment.color.a == 0) discard;
 
     // calculate normal from texture coordinates
     vec3 normal;
     normal.xy = gl_PointCoord * vec2(2.0, -2.0) + vec2(-1.0, 1.0);
     float rad = sqrt(dot(normal.xy, normal.xy));
-    if (rad > 1.0) {
-        discard;  // kill pixels outside circle
-    }
+    if (rad > 1.0)  discard;  // kill pixels outside circle
 
     // Get linear depth
     float z_v = convertDepthScreenToView(camera, gl_FragCoord.z);  // view space depth
-    float depth =
-        (z_v - camera.nearPlane) / (camera.farPlane - camera.nearPlane);  // linear normalised depth
+    // linear normalised depth
+    float depth = (z_v - camera.nearPlane) / (camera.farPlane - camera.nearPlane);
 
     // Calculate g_i^2
 #ifdef USE_IMPORTANCE_VOLUME
@@ -89,21 +80,15 @@ void main() {
 #endif
 
     // Project importance
-    float gisq = gi * gi;
-    float gtot = total(importanceSumCoeffs[0], N_IMPORTANCE_SUM_COEFFICIENTS);
-    float Gd = approximate(importanceSumCoeffs[0], N_IMPORTANCE_SUM_COEFFICIENTS, depth);
-    Gd += 0.5 * gisq;  // correct for importance sum approximation at discontinuity
-    float alpha =
-        clamp(1 / (1 + pow(1 - gi, 2 * lambda) * (r * max(0, Gd - gisq) + q * max(0, gtot - Gd))),
-              0.0, 0.9999);  // set pixel alpha using opacity optimisation
+    float alpha = projectImportance(gi, depth, importanceSumCoeffs[0], N_IMPORTANCE_SUM_COEFFICIENTS);
 
     float glyphRadius = pointSize * 0.5;
     rad *= pointSize * 0.5 + borderWidth;
 
     // pseudo antialiasing with the help of the alpha channel
     // i.e. smooth transition between center and border, and smooth alpha fall-off at the outer rim
-    float outerglyphRadius = glyphRadius + borderWidth -
-                             antialising;  // used for adjusting the alpha value of the outer rim
+    // used for adjusting the alpha value of the outer rim
+    float outerglyphRadius = glyphRadius + borderWidth - antialising;
 
     float borderAlpha = clamp(
         mix(1.0, 0.0, (rad - outerglyphRadius) / (glyphRadius + borderWidth - outerglyphRadius)),
